lots of changes to get runti working properly
This is quite a commit. There is a lot of stuff going on in here.

First, both subprocesses (chef-client and runit) are now launched simply as an exec inside a fork. There was some weirdness with Process.spawn using a subshell which was giving me back the wrong child pid so I couldn't accurately reap it. The Open3 library also seemed like overkill upon further inspection since I was using it primarily to stream the logs to stdout but I realized that running it in exec would accomplish the same thing.

Next, I reworked how I kill all the processes. There was some confusion around what <defunct> processes are but I feel like I've got those all taken care of. chef-init will now reliably and safely kill all the subprocesses.

Once I did that, I was able to write clear verify tests that will work as expected. I also now check to make sure that processes are cleaned up after every run.

The last thing I did was update the spec tests for style, formatting and with some of the changes I made to cli.rb and verify.rb