Add ChefInit::Process; Improve testing framework
To better handle processes inside chef-init, I created a ChefInit::Process class that takes advantage of the sys-proctable gem to manage processes. This cleans things up in chef-init/cli and chef-init/verify.

I also added methods for different types of tests: test_ps, test_cmd and test_file. This reduces a lot of the duplication that was occuring.