Use ChildProcess Gem
Instead of a custom Process class that tries to grok status from the process table, we are going to use the childprocess gem which is a cross-platform implementation of the custom class we were working on.

Since we want to block until the chef-client has finished, I am going to use standard system_command (Mixlib::Shellout) for the chef-client run.