Pretty much a complete rewrite
We are now leveraging the Chef::Application library instead of creating our own pattern for handling the running process. As such some of the code has been moved into the chef namespace.

We are also removing dependencies on external gems other than Chef.

Also, we are removing the direct use of embedded runit and instead requiring that the runit binaries runsvdir, runsv and sv be in the path. This, along with the supervisor process command being a configurable value, sets the ground for the possibility of using your own supervisor.